Function to perform dimension reduction

This function takes the distance matrix as input and reduce the dimension to 2 axis

Usage

dim_reduction(distance_matrix)

Arguments

distance_matrix

distance matrix

Value

A dataframe containing the id, position in x axis, position in y axis

Examples

dim_reduction(distance_matrix)
#>                          axis_1       axis_2
#> TcYC6_0128200-RA-p1 -3.75429634 -1.304527740
#> TcYC6_0128210-RA-p1 -3.11087130 -0.691658777
#> TcYC6_0128220-RA-p1 -4.13674651 -1.698944011
#> TcYC6_0128230-RA-p1 -3.82431872 -1.540161121
#> TcYC6_0128240-RA-p1 -4.32919109 -1.648458709
#> TcYC6_0128250-RA-p1 -4.18799408 -1.730389350
#> TcYC6_0128260-RA-p1 -4.29841906 -1.595174655
#> TcYC6_0128270-RA-p1 -4.15979475 -1.663530325
#> TcYC6_0128280-RA-p1 -4.27965990 -1.549149274
#> TcYC6_0128290-RA-p1 -4.03518989 -1.566746578
#> TcYC6_0128300-RA-p1 -4.31998295 -1.630271813
#> TcYC6_0128310-RA-p1 -4.12136567 -1.673768282
#> TcYC6_0128320-RA-p1 -4.13682508 -1.655352691
#> TcYC6_0128330-RA-p1 -3.97206330 -1.529543175
#> TcYC6_0128340-RA-p1 -3.19856040 -1.051077946
#> TcYC6_0128350-RA-p1 -2.59738806 -0.583696758
#> TcYC6_0128360-RA-p1 -4.12023996 -1.606423931
#> TcYC6_0128370-RA-p1 -4.02573070 -1.575551442
#> TcYC6_0128380-RA-p1 -3.16098060 -0.942651847
#> TcYC6_0128390-RA-p1 -3.14934120 -0.947986466
#> TcYC6_0128400-RA-p1 -3.18784285 -0.937280105
#> TcYC6_0128410-RA-p1 -3.15579457 -1.012094244
#> TcYC6_0128420-RA-p1 -3.20412653 -0.997276080
#> TcYC6_0128430-RA-p1 -3.18768863 -0.999765743
#> TcYC6_0128440-RA-p1 -3.20308982 -0.989154550
#> TcYC6_0128450-RA-p1 -3.20585087 -0.995754829
#> TcYC6_0128460-RA-p1 -3.96175391 -1.490067837
#> TcYC6_0128470-RA-p1 -3.93825998 -1.531764939
#> TcYC6_0128480-RA-p1 -3.87343437 -1.514486949
#> TcYC6_0128490-RA-p1 -4.05661436 -1.612759828
#> TcYC6_0128500-RA-p1 -3.88068459 -1.443443002
#> TcYC6_0128510-RA-p1 -3.87415588 -1.540276565
#> TcYC6_0135710-RA-p1  1.18446589  0.201849833
#> TcYC6_0135720-RA-p1 -0.28307965  6.968338122
#> TcYC6_0135730-RA-p1 -0.30861764  6.966791222
#> TcYC6_0135740-RA-p1 -0.28257768  6.959733051
#> TcYC6_0137830-RA-p1 10.20186417 -1.614151352
#> TcYC6_0137840-RA-p1  9.98499630 -1.442802336
#> TcYC6_0137850-RA-p1 10.09063452 -1.419978040
#> TcYC6_0137860-RA-p1  9.59720567 -1.429954530
#> TcYC6_0137870-RA-p1  9.94576317 -1.394069771
#> TcYC6_0137880-RA-p1 10.04105796 -1.471683771
#> TcYC6_0137890-RA-p1 10.08586287 -1.585244803
#> TcYC6_0137900-RA-p1  9.91174039 -1.549077665
#> TcYC6_0130340-RA-p1 -4.12004055 -1.406881441
#> TcYC6_0130350-RA-p1 -0.48910875 -0.162454010
#> TcYC6_0130360-RA-p1 -2.74643612 -0.621955463
#> TcYC6_0130370-RA-p1 -2.75232061 -0.608241615
#> TcYC6_0130380-RA-p1 -2.73123740 -0.616869598
#> TcYC6_0130390-RA-p1 -4.10405012 -1.403172508
#> TcYC6_0136230-RA-p1 -0.18241015  6.657374071
#> TcYC6_0131560-RA-p1 -4.08782917 -1.549505652
#> TcYC6_0131570-RA-p1 -3.96490886 -1.482491137
#> TcYC6_0131580-RA-p1 -4.01132992 -1.296903899
#> TcYC6_0131590-RA-p1 -2.94716020 -0.842965545
#> TcYC6_0131600-RA-p1 -3.94337730 -1.626155423
#> TcYC6_0131610-RA-p1 -3.95055076 -1.615574596
#> TcYC6_0131620-RA-p1 -4.05909276 -1.342952089
#> TcYC6_0131630-RA-p1 -4.14291860 -1.453922453
#> TcYC6_0131640-RA-p1 -3.96890917 -1.641268449
#> TcYC6_0131650-RA-p1 -4.26880814 -1.514782396
#> TcYC6_0131660-RA-p1 -3.99252683 -1.605711310
#> TcYC6_0131670-RA-p1 -4.27273733 -1.533986783
#> TcYC6_0131680-RA-p1 -3.95242526 -1.599473404
#> TcYC6_0131690-RA-p1 -4.13830831 -1.451218467
#> TcYC6_0131700-RA-p1 -2.88759680 -1.002929793
#> TcYC6_0131710-RA-p1 -3.83750768 -1.485475462
#> TcYC6_0131720-RA-p1 -3.94841838 -1.625064113
#> TcYC6_0131730-RA-p1 -3.96309151 -1.598535906
#> TcYC6_0131740-RA-p1 -3.97265043 -1.623048999
#> TcYC6_0131750-RA-p1 -3.97265043 -1.623048999
#> TcYC6_0131760-RA-p1 -3.97265043 -1.623048999
#> TcYC6_0131770-RA-p1 -3.97411459 -1.623570271
#> TcYC6_0131780-RA-p1 -3.19079695 -0.984369520
#> TcYC6_0131790-RA-p1 -4.04360477 -1.592833377
#> TcYC6_0131800-RA-p1 -4.03648466 -1.543619569
#> TcYC6_0131810-RA-p1 -1.90331376 -0.834998025
#> TcYC6_0136630-RA-p1 -3.49303881 -1.411068411
#> TcYC6_0136640-RA-p1 -0.29073407  6.861449707
#> TcYC6_0136650-RA-p1 -0.20204736  6.842909512
#> TcYC6_0138970-RA-p1  9.19353859 -1.209469100
#> TcYC6_0138980-RA-p1  9.21193259 -1.272820135
#> TcYC6_0138990-RA-p1 10.06399173 -0.988703558
#> TcYC6_0139000-RA-p1  3.80610612  0.122465053
#> TcYC6_0121890-RA-p1  2.90592975  0.720094875
#> TcYC6_0123350-RA-p1  2.97742372  0.615847073
#> TcYC6_0131830-RA-p1 -3.12378831 -0.912848944
#> TcYC6_0131840-RA-p1 -2.34802958 -0.224173267
#> TcYC6_0131850-RA-p1 -1.42201087 -0.006446663
#> TcYC6_0139020-RA-p1  6.28162750 -0.933692920
#> TcYC6_0139030-RA-p1  6.30971800 -0.946515730
#> TcYC6_0006870-RA-p1  3.33522706  0.296496566
#> TcYC6_0130940-RA-p1 -4.05864707 -1.568021549
#> TcYC6_0130950-RA-p1 -4.06930834 -1.699227045
#> TcYC6_0130960-RA-p1 -2.88626094 -0.904548978
#> TcYC6_0130970-RA-p1 -2.87363758 -0.932954608
#> TcYC6_0130980-RA-p1 -2.87366284 -0.939683074
#> TcYC6_0130990-RA-p1 -2.83882047 -0.898146891
#> TcYC6_0131000-RA-p1 -2.83882047 -0.898146891
#> TcYC6_0131010-RA-p1 -2.86252401 -0.915162395
#> TcYC6_0131020-RA-p1 -3.27059782 -1.048550793
#> TcYC6_0131030-RA-p1 -4.22680212 -1.543421015
#> TcYC6_0131040-RA-p1 -3.10837645 -0.644624910
#> TcYC6_0131050-RA-p1 -3.06084796 -0.984958711
#> TcYC6_0131060-RA-p1 -4.09720014 -1.708836239
#> TcYC6_0131070-RA-p1 -3.30241245 -0.981911155
#> TcYC6_0131080-RA-p1 -4.12780254 -1.683273328
#> TcYC6_0131090-RA-p1 -4.20681358 -1.538063877
#> TcYC6_0131100-RA-p1 -4.10856086 -1.709942811
#> TcYC6_0131110-RA-p1 -4.28447926 -1.548104286
#> TcYC6_0131120-RA-p1 -4.21958706 -1.392141091
#> TcYC6_0131130-RA-p1 -4.11547386 -1.412228639
#> TcYC6_0136420-RA-p1 -3.91030320 -1.628647275
#> TcYC6_0136430-RA-p1 -3.91030320 -1.628647275
#> TcYC6_0136440-RA-p1  0.42121460 -0.505555666
#> TcYC6_0136450-RA-p1 -1.45713899  0.089873821
#> TcYC6_0136460-RA-p1 -0.19340273  6.838049493
#> TcYC6_0138630-RA-p1  6.38043842 -1.079019513
#> TcYC6_0138640-RA-p1 10.28395880 -1.558832402
#> TcYC6_0138650-RA-p1  7.15096182  0.182504979
#> TcYC6_0138660-RA-p1  7.26607467  0.140339153
#> TcYC6_0138670-RA-p1 10.04532996 -1.506276002
#> TcYC6_0138680-RA-p1  3.81269322  0.146373136
#> TcYC6_0138690-RA-p1  3.81587517  0.159699902
#> TcYC6_0058000-RA-p1  3.56231975  0.565337829
#> TcYC6_0059120-RA-p1  2.98185928  0.613735279
#> TcYC6_0129710-RA-p1 -4.16520902 -1.366267614
#> TcYC6_0129720-RA-p1 -2.77620167 -0.853699392
#> TcYC6_0129730-RA-p1 -4.14228167 -1.397025015
#> TcYC6_0129740-RA-p1 -4.06285503 -1.431480598
#> TcYC6_0129750-RA-p1 -4.02172468 -1.291635426
#> TcYC6_0129760-RA-p1 -4.16086731 -1.364781821
#> TcYC6_0129770-RA-p1 -4.16823906 -1.385748290
#> TcYC6_0129780-RA-p1 -4.21391233 -1.512959649
#> TcYC6_0129790-RA-p1  3.76740067  0.295116130
#> TcYC6_0129800-RA-p1 -3.99172409 -1.644049102
#> TcYC6_0129810-RA-p1 -4.13090927 -1.569707789
#> TcYC6_0129820-RA-p1 -4.18214587 -1.382162024
#> TcYC6_0129830-RA-p1 -4.13337681 -1.409197390
#> TcYC6_0136040-RA-p1 -2.07087444 -0.512420809
#> TcYC6_0136050-RA-p1 -2.05654681 -0.528683207
#> TcYC6_0136060-RA-p1 -2.08589565 -0.539189227
#> TcYC6_0136070-RA-p1 -2.06954157 -0.530010767
#> TcYC6_0136080-RA-p1 -2.06954157 -0.530010767
#> TcYC6_0138270-RA-p1 10.04238605 -1.318468462
#> TcYC6_0138280-RA-p1  8.26621309 -1.461055438
#> TcYC6_0138290-RA-p1 10.04238605 -1.318468462
#> TcYC6_0138300-RA-p1  9.61859942 -1.287255061
#> TcYC6_0130930-RA-p1 -2.73088003 -0.711311076
#> TcYC6_0136320-RA-p1 -0.30112420  6.994196668
#> TcYC6_0136330-RA-p1 -0.29432530  6.986552884
#> TcYC6_0136340-RA-p1 -0.30588827  6.973089597
#> TcYC6_0136350-RA-p1 -0.29432530  6.986552884
#> TcYC6_0136360-RA-p1 -0.29393912  6.978167123
#> TcYC6_0136370-RA-p1 -0.30328126  6.988266447
#> TcYC6_0136380-RA-p1 -0.30112420  6.994196668
#> TcYC6_0136390-RA-p1 -0.30112420  6.994196668
#> TcYC6_0136400-RA-p1 -0.29308596  6.972573777
#> TcYC6_0136410-RA-p1 -0.29343197  6.954898341
#> TcYC6_0138600-RA-p1  9.88199098 -1.422126267
#> TcYC6_0131480-RA-p1 -3.03060340 -0.965699930
#> TcYC6_0131490-RA-p1 -2.35059603 -0.258105123
#> TcYC6_0138960-RA-p1  9.61859942 -1.287255061
#> TcYC6_0131500-RA-p1 -3.28679956 -1.034022673
#> TcYC6_0131510-RA-p1 -3.29836326 -1.049175434
#> TcYC6_0131520-RA-p1 -3.25915617 -1.023670627
#> TcYC6_0131530-RA-p1 -3.26467238 -1.018682130
#> TcYC6_0131540-RA-p1 -2.55010753 -0.834803969
#> TcYC6_0107720-RA-p1  2.91501384  0.570231931
#> TcYC6_0131230-RA-p1 -2.88766851 -0.987533216
#> TcYC6_0131240-RA-p1 -3.25113353 -0.954229462
#> TcYC6_0131250-RA-p1  3.04591336  0.203435665
#> TcYC6_0131260-RA-p1  2.99644500  0.177087732
#> TcYC6_0138810-RA-p1  9.97407587 -1.212664001
#> TcYC6_0138820-RA-p1  9.95548290 -1.317151436
#> TcYC6_0138830-RA-p1  8.77200024 -1.410340625
#> TcYC6_0138840-RA-p1 10.31619572 -1.604156019
#> TcYC6_0135960-RA-p1 -0.27865058  6.933802907
#> TcYC6_0135970-RA-p1 -0.27818814  6.929493310
#> TcYC6_0135980-RA-p1 -0.27562754  6.911222086
#> TcYC6_0135990-RA-p1 -0.19616579  6.826155913
#> TcYC6_0136000-RA-p1 -0.17825902  6.784332721
#> TcYC6_0139060-RA-p1  9.78279424 -1.445227302
#> TcYC6_0139070-RA-p1  9.73002563 -1.491156050
#> TcYC6_0139080-RA-p1  9.73002563 -1.491156050
#> TcYC6_0131140-RA-p1 -3.26437365 -1.011780948
#> TcYC6_0131150-RA-p1 -3.16527970 -0.595950872
#> TcYC6_0131160-RA-p1 -3.26239251 -0.958774683
#> TcYC6_0131170-RA-p1 -3.97657996 -1.556674542
#> TcYC6_0131180-RA-p1 -3.99698488 -1.524487194
#> TcYC6_0136470-RA-p1  0.02081438  6.249624688
#> TcYC6_0136480-RA-p1 -0.16543571  6.641274450
#> TcYC6_0138700-RA-p1  9.83068002 -1.474711050
#> TcYC6_0138710-RA-p1  9.27056034 -1.265330738
#> TcYC6_0086470-RA-p1  3.37322339  0.492376073
#> TcYC6_0130820-RA-p1 -4.23240738 -1.608743409
#> TcYC6_0130830-RA-p1 -4.26806282 -1.573710511
#> TcYC6_0130840-RA-p1 -4.07331314 -1.752225947
#> TcYC6_0130850-RA-p1 -3.13704528 -0.943708937
#> TcYC6_0130860-RA-p1 -2.59574531 -0.931546856
#> TcYC6_0130870-RA-p1 -2.79244407 -0.940945880
#> TcYC6_0130880-RA-p1 -4.23923630 -1.587997627
#> TcYC6_0130890-RA-p1 -4.27900026 -1.537384932
#> TcYC6_0130900-RA-p1 -1.75616890  0.064718432
#> TcYC6_0130910-RA-p1  0.99512684  0.084835622
#> TcYC6_0130920-RA-p1 -1.22992015  0.337179222
#> TcYC6_0130460-RA-p1 -3.29717366 -1.073663061
#> TcYC6_0130470-RA-p1 -2.76431329 -0.586051705
#> TcYC6_0130480-RA-p1 -2.75618286 -0.581863662
#> TcYC6_0130490-RA-p1 -2.75511570 -0.582354159
#> TcYC6_0130500-RA-p1 -2.76404409 -0.884866713
#> TcYC6_0130510-RA-p1 -2.76564877 -0.919794076
#> TcYC6_0130520-RA-p1 -4.23403997 -1.470349255
#> TcYC6_0130530-RA-p1 -3.96447720 -1.516662685
#> TcYC6_0130540-RA-p1 -2.76484659 -0.596399312
#> TcYC6_0136250-RA-p1 -0.29775741  6.974527860
#> TcYC6_0136260-RA-p1 -0.02517578  6.492353606
#> TcYC6_0138500-RA-p1 10.11222191 -1.450286656
#> TcYC6_0138510-RA-p1  9.60705480 -1.286594247
#> TcYC6_0138520-RA-p1 10.04553069 -1.473425217
#> TcYC6_0138530-RA-p1  8.95160004 -1.301291572
#> TcYC6_0138540-RA-p1  9.58732517 -1.578790703
#> TcYC6_0138550-RA-p1 10.11222191 -1.450286656
#> TcYC6_0137970-RA-p1  9.30252648 -1.610568341
#> TcYC6_0137980-RA-p1  6.61475158  0.131023608
#> TcYC6_0130090-RA-p1 -3.23957272 -1.016951337
#> TcYC6_0130100-RA-p1 -2.64948468 -0.866121861
#> TcYC6_0130110-RA-p1 -3.03403600 -0.971488207
#> TcYC6_0136150-RA-p1 -0.29343197  6.954898341
#> TcYC6_0138440-RA-p1  9.72527798 -1.363413042
#> TcYC6_0131910-RA-p1 -3.02009457 -0.844265561
#> TcYC6_0131920-RA-p1 -2.70810459 -0.598458344
#> TcYC6_0131930-RA-p1 -2.98631471 -0.941367791
#> TcYC6_0131940-RA-p1 -3.00462584 -0.948821970
#> TcYC6_0131950-RA-p1 -3.13847684 -0.650170293
#> TcYC6_0131960-RA-p1 -3.06862692 -0.897918321
#> TcYC6_0131970-RA-p1 -3.04044648 -0.943072342
#> TcYC6_0136660-RA-p1 -0.29432530  6.986552884
#> TcYC6_0136670-RA-p1 -0.29432530  6.986552884
#> TcYC6_0136680-RA-p1 -0.29019804  6.972747651
#> TcYC6_0136690-RA-p1 -0.30280836  6.965835725
#> TcYC6_0136700-RA-p1 -0.29343197  6.954898341
#> TcYC6_0130730-RA-p1 -3.07810915 -0.604101359
#> TcYC6_0136310-RA-p1 -0.18553735  6.733362641
#> TcYC6_0025940-RA-p1  3.49729071  0.448713051
#> TcYC6_0025960-RA-p1  3.50138089  0.459403330
#> TcYC6_0128700-RA-p1 -3.71380881 -1.452175318
#> TcYC6_0128710-RA-p1 -2.99671020 -0.877272256
#> TcYC6_0128720-RA-p1 -3.01493526 -0.865406058
#> TcYC6_0128730-RA-p1 -2.99671020 -0.877272256
#> TcYC6_0128740-RA-p1 -2.58812569 -0.847253524
#> TcYC6_0128750-RA-p1 -3.21030886 -0.701556587
#> TcYC6_0128760-RA-p1 -3.20475492 -0.726539043
#> TcYC6_0128770-RA-p1 -3.18831759 -0.701492716
#> TcYC6_0128780-RA-p1 -3.18797005 -0.720259105
#> TcYC6_0128790-RA-p1 -3.22214134 -0.713468533
#> TcYC6_0128800-RA-p1 -3.21572117 -0.708409683
#> TcYC6_0128810-RA-p1 -3.17878579 -0.713540767
#> TcYC6_0128820-RA-p1 -2.83641119 -0.898953497
#> TcYC6_0128830-RA-p1 -2.82760699 -0.906052968
#> TcYC6_0128840-RA-p1 -2.83377212 -0.908452291
#> TcYC6_0128850-RA-p1 -4.15887620 -1.391385226
#> TcYC6_0128860-RA-p1 -2.84575497 -0.853851464
#> TcYC6_0135880-RA-p1 -0.29774926  6.997399289
#> TcYC6_0130240-RA-p1 -2.00780304 -0.606793675
#> TcYC6_0130250-RA-p1 -4.20680893 -1.447151433
#> TcYC6_0130260-RA-p1 -3.04684250 -0.928762123
#> TcYC6_0130270-RA-p1  1.79369147 -0.434270377
#> TcYC6_0130280-RA-p1 -2.21974902 -0.788332262
#> TcYC6_0138470-RA-p1 10.45050935 -1.532547120
#> TcYC6_0138480-RA-p1  6.56749534 -1.179810379
#> TcYC6_0129540-RA-p1 -2.34801186 -0.824490551
#> TcYC6_0129550-RA-p1 -4.21764225 -1.495049797
#> TcYC6_0129560-RA-p1 -4.15154354 -1.445395267
#> TcYC6_0129570-RA-p1 -4.14481594 -1.483869171
#> TcYC6_0129580-RA-p1 -4.27165812 -1.517046398
#> TcYC6_0129590-RA-p1 -4.27934915 -1.468920193
#> TcYC6_0129600-RA-p1 -4.19352996 -1.450164719
#> TcYC6_0129610-RA-p1 -4.18328354 -1.497048705
#> TcYC6_0129620-RA-p1  0.02292839  5.965095695
#> TcYC6_0136030-RA-p1 -4.16742369 -1.444354495
#> TcYC6_0079290-RA-p1  3.07003303  0.575447060
#> TcYC6_0079380-RA-p1  3.08375486  0.589494844
#> TcYC6_0080760-RA-p1 -0.12674897  0.640833247
#> TcYC6_0130650-RA-p1 -4.03405334 -1.522276043
#> TcYC6_0130660-RA-p1 -4.08552824 -1.548520783
#> TcYC6_0130670-RA-p1 -4.10641500 -1.275303803
#> TcYC6_0130680-RA-p1 -3.97852670 -1.476024534
#> TcYC6_0130690-RA-p1 -2.93290660 -0.731389251
#> TcYC6_0136300-RA-p1 -3.08311586 -0.615183516
#> TcYC6_0131190-RA-p1 -4.05269739 -1.493515932
#> TcYC6_0131200-RA-p1 -3.89480262 -1.487483716
#> TcYC6_0131210-RA-p1 -3.10139309 -0.610624254
#> TcYC6_0131220-RA-p1 -2.70767574 -0.588326489
#> TcYC6_0136510-RA-p1 -0.30112420  6.994196668
#> TcYC6_0138750-RA-p1  9.70428970 -0.953064187
#> TcYC6_0138760-RA-p1  9.60966351 -1.567309492
#> TcYC6_0130160-RA-p1 -3.27590092 -1.043912157
#> TcYC6_0130170-RA-p1 -2.06301757 -0.685787520
#> TcYC6_0130180-RA-p1 -2.06987888 -0.689884386
#> TcYC6_0130190-RA-p1 -3.18799515 -0.662474486
#> TcYC6_0136190-RA-p1 -0.20452885  6.821342524
#> TcYC6_0138460-RA-p1  9.65677064 -1.217744493
#> TcYC6_0048160-RA-p1  3.84284655  0.556307116
#> TcYC6_0129340-RA-p1 -3.18594118 -0.689704731
#> TcYC6_0129350-RA-p1 -3.11792365 -0.681244226
#> TcYC6_0129360-RA-p1 -3.11934797 -0.717082416
#> TcYC6_0129370-RA-p1 -2.61479345 -0.931998651
#> TcYC6_0129380-RA-p1 -2.61479345 -0.931998651
#> TcYC6_0129390-RA-p1 -2.61295006 -0.925217065
#> TcYC6_0129400-RA-p1 -2.90100186 -0.989225053
#> TcYC6_0129410-RA-p1 -2.92109549 -0.958942563
#> TcYC6_0129420-RA-p1 -2.93942840 -0.942609540
#> TcYC6_0129430-RA-p1 -3.21711621 -0.720747145
#> TcYC6_0129440-RA-p1 -3.22790155 -0.728975733
#> TcYC6_0129450-RA-p1 -3.21157183 -0.722950013
#> TcYC6_0129460-RA-p1 -3.15882132 -0.664041162
#> TcYC6_0129470-RA-p1 -3.14299200 -0.708334767
#> TcYC6_0129480-RA-p1 -3.17781217 -0.717834221
#> TcYC6_0129490-RA-p1  3.31851455  0.254860035
#> TcYC6_0129510-RA-p1  3.68295821  0.179410990
#> TcYC6_0135950-RA-p1 -0.18905758  6.750520489
#> TcYC6_0138230-RA-p1  9.43286975 -0.732041944
#> TcYC6_0130770-RA-p1 -2.62509879 -0.355460475
#> TcYC6_0130780-RA-p1 -2.64103821 -0.366673354
#> TcYC6_0130790-RA-p1 -2.65344274 -0.381434744
#> TcYC6_0130800-RA-p1 -2.65999364 -0.360580460
#> TcYC6_0130810-RA-p1 -2.63722539 -0.363107042
#> TcYC6_0138580-RA-p1  9.63095209 -1.173512284
#> TcYC6_0138590-RA-p1  5.19518255  1.806609563
#> TcYC6_0136490-RA-p1 -2.79201738 -0.564021885
#> TcYC6_0136500-RA-p1 -0.29774926  6.997399289
#> TcYC6_0138720-RA-p1  9.84473873 -1.399228184
#> TcYC6_0138730-RA-p1  9.72069401 -1.479318102
#> TcYC6_0138740-RA-p1  9.72069401 -1.479318102
#> TcYC6_0131290-RA-p1 -2.57062223 -0.927662791
#> TcYC6_0131300-RA-p1 -2.21860135 -0.793661552
#> TcYC6_0131310-RA-p1 -4.12618687 -1.519342166
#> TcYC6_0131320-RA-p1 -2.21370756 -0.808746684
#> TcYC6_0131330-RA-p1 -2.28708589 -0.777794519
#> TcYC6_0131340-RA-p1 -2.27035494 -0.809040068
#> TcYC6_0131350-RA-p1 -2.21394669 -0.721511114
#> TcYC6_0131360-RA-p1 -2.79023217 -0.538897823
#> TcYC6_0131370-RA-p1 -2.75199154 -0.526674081
#> TcYC6_0131390-RA-p1  3.65995417  0.183647263
#> TcYC6_0136530-RA-p1 -0.29260524  6.982966685
#> TcYC6_0136540-RA-p1 -0.21249439  6.883036630
#> TcYC6_0136550-RA-p1 -0.12720313  6.624754608
#> TcYC6_0138850-RA-p1 10.44870852 -1.579504383
#> TcYC6_0138860-RA-p1  9.56258950 -1.491140416
#> TcYC6_0138870-RA-p1  9.94953763 -1.401612175
#> TcYC6_0138880-RA-p1  9.46519370 -1.400171251
#> TcYC6_0138890-RA-p1  9.55209828 -1.449039703
#> TcYC6_0138900-RA-p1 10.15970341 -1.390542743
#> TcYC6_0138910-RA-p1 10.34011656 -1.457370263
#> TcYC6_0138920-RA-p1 10.15504192 -1.332664871
#> TcYC6_0138930-RA-p1  9.51603327 -1.454987276
#> TcYC6_0138940-RA-p1  9.47139164 -1.338432371
#> TcYC6_0138950-RA-p1 10.73908575 -1.139861322
#> TcYC6_0129230-RA-p1 -3.88318509 -1.477500561
#> TcYC6_0129240-RA-p1 -4.28397518 -1.544698713
#> TcYC6_0129250-RA-p1 -4.22963745 -1.427660457
#> TcYC6_0129260-RA-p1 -3.24584919 -1.149741897
#> TcYC6_0129270-RA-p1 -3.09608153 -1.068887024
#> TcYC6_0129280-RA-p1 -3.11153174 -1.092154562
#> TcYC6_0129290-RA-p1 -3.14344227 -1.144178294
#> TcYC6_0129300-RA-p1  7.21395360  0.105753608
#> TcYC6_0135930-RA-p1 -4.19460975 -1.476065055
#> TcYC6_0135940-RA-p1 -3.11177382 -1.106247759
#> TcYC6_0138120-RA-p1  9.67148989 -1.386632380
#> TcYC6_0138130-RA-p1 10.16665311 -1.297761928
#> TcYC6_0138140-RA-p1 10.81599549 -1.235139368
#> TcYC6_0138150-RA-p1  9.95486903 -1.288890254
#> TcYC6_0138160-RA-p1 10.79714801 -1.219245938
#> TcYC6_0138170-RA-p1 10.79714801 -1.219245938
#> TcYC6_0138180-RA-p1  9.91621233 -1.264017479
#> TcYC6_0138190-RA-p1 10.77342737 -1.158048702
#> TcYC6_0138200-RA-p1 10.80008956 -1.153554853
#> TcYC6_0138210-RA-p1  9.93547589 -1.216623362
#> TcYC6_0128570-RA-p1 -4.05675702 -1.476980819
#> TcYC6_0135850-RA-p1 -3.81475864 -1.352072666
#> TcYC6_0137930-RA-p1  5.96318360 -0.807730689
#> TcYC6_0137940-RA-p1  7.59302622 -0.050646361
#> TcYC6_0137950-RA-p1  7.45233411 -0.057759892
#> TcYC6_0131400-RA-p1 -3.10172152 -0.608262680
#> TcYC6_0131410-RA-p1 -2.19699960 -0.683446010
#> TcYC6_0131420-RA-p1 -2.02419767 -0.517525284
#> TcYC6_0131430-RA-p1 -2.21050656 -0.742105138
#> TcYC6_0131440-RA-p1 -2.80880476 -0.962074397
#> TcYC6_0131450-RA-p1 -2.80880476 -0.962074397
#> TcYC6_0131460-RA-p1 -1.63000726 -0.269848371
#> TcYC6_0128920-RA-p1 -4.11571865 -1.464187146
#> TcYC6_0128930-RA-p1 -2.73579537 -0.893744899
#> TcYC6_0128940-RA-p1 -4.05859637 -1.489469401
#> TcYC6_0128950-RA-p1 -4.05340630 -1.586615812
#> TcYC6_0128960-RA-p1 -4.06789649 -1.591383664
#> TcYC6_0128970-RA-p1  3.65175679  0.205151348
#> TcYC6_0128980-RA-p1 -4.17303307 -1.624575306
#> TcYC6_0128990-RA-p1 -2.52322463 -0.926173998
#> TcYC6_0129000-RA-p1 -2.74297198 -0.468462629
#> TcYC6_0129010-RA-p1 -4.22043179 -1.616865320
#> TcYC6_0129020-RA-p1 -3.96738265 -1.273105838
#> TcYC6_0129030-RA-p1 -4.05205485 -1.567299335
#> TcYC6_0129040-RA-p1 -4.07369036 -1.522130330
#> TcYC6_0129050-RA-p1 -4.15729614 -1.570306958
#> TcYC6_0129060-RA-p1 -4.15729614 -1.570306958
#> TcYC6_0129070-RA-p1  1.05304465  0.109893672
#> TcYC6_0135900-RA-p1 -2.54898101 -1.290344498
#> TcYC6_0135910-RA-p1 -0.01804264  6.402381781
#> TcYC6_0138020-RA-p1  8.12016999 -1.082158221
#> TcYC6_0138030-RA-p1 10.73908575 -1.139861322
#> TcYC6_0138040-RA-p1 10.04636018 -1.418639743
#> TcYC6_0138050-RA-p1 10.04636018 -1.418639743
#> TcYC6_0138060-RA-p1 10.04636018 -1.418639743
#> TcYC6_0138610-RA-p1 10.54792998 -1.341189683
#> TcYC6_0138620-RA-p1  5.62310104 -1.087246135
#> TcYC6_0060660-RA-p1  3.94270657  0.510775971
#> TcYC6_0061820-RA-p1  3.95178470  0.557698193
#> TcYC6_0129890-RA-p1 -2.91851592 -1.017263763
#> TcYC6_0129900-RA-p1 -2.82682817 -0.570851185
#> TcYC6_0129910-RA-p1 -3.28431298 -1.144499372
#> TcYC6_0129920-RA-p1 -3.84713835 -1.440235053
#> TcYC6_0129930-RA-p1 -2.58255678 -1.003308452
#> TcYC6_0129940-RA-p1 -2.81204324 -0.550226370
#> TcYC6_0129950-RA-p1 -2.86502833 -0.538827618
#> TcYC6_0138330-RA-p1  9.88830664 -1.443770234
#> TcYC6_0138340-RA-p1  9.57346416 -1.037770414
#> TcYC6_0138350-RA-p1  9.71591437 -1.293118802
#> TcYC6_0138360-RA-p1  9.66997961 -1.300539126
#> TcYC6_0061980-RA-p1  3.26905923  0.643610817
#> TcYC6_0129960-RA-p1 -4.24245059 -1.481769532
#> TcYC6_0129970-RA-p1 -3.93132724 -1.589500280
#> TcYC6_0129980-RA-p1 -3.96559917 -1.590252575
#> TcYC6_0129990-RA-p1 -4.22233040 -1.470624801
#> TcYC6_0130000-RA-p1 -4.22448409 -1.492185538
#> TcYC6_0130010-RA-p1 -2.92216279 -1.028998992
#> TcYC6_0130020-RA-p1 -4.26450782 -1.457223059
#> TcYC6_0130030-RA-p1 -4.25966019 -1.463227927
#> TcYC6_0130040-RA-p1 -4.30932273 -1.440813751
#> TcYC6_0130050-RA-p1 -4.17744165 -1.525209632
#> TcYC6_0130060-RA-p1 -4.23708572 -1.553608241
#> TcYC6_0130070-RA-p1 -4.23285871 -1.402549941
#> TcYC6_0130080-RA-p1 -4.34748143 -1.459912154
#> TcYC6_0136100-RA-p1 -1.48539464 -0.145806676
#> TcYC6_0136110-RA-p1 -1.50139811 -0.122268865
#> TcYC6_0136120-RA-p1 -1.56928106 -0.036171821
#> TcYC6_0136130-RA-p1 -1.55167289  0.025134765
#> TcYC6_0136140-RA-p1  0.53578347  5.569189681
#> TcYC6_0138370-RA-p1 10.19274853 -1.558953099
#> TcYC6_0138380-RA-p1  9.94106988 -1.448088501
#> TcYC6_0138390-RA-p1  9.83989893 -1.322162395
#> TcYC6_0138400-RA-p1  9.75694573 -1.269911515
#> TcYC6_0138410-RA-p1  9.80518278 -1.251020291
#> TcYC6_0138420-RA-p1  7.23738093  0.146358056
#> TcYC6_0138430-RA-p1  7.26798288  0.061890340
#> TcYC6_0128640-RA-p1 -4.17350862 -1.467225878
#> TcYC6_0135870-RA-p1 -0.06124065  6.550046868
#> TcYC6_0138790-RA-p1  8.00770831 -0.282903761
#> TcYC6_0138800-RA-p1  3.69234957  0.172800811
#> TcYC6_0131880-RA-p1 -4.23661182 -1.463921633
#> TcYC6_0131890-RA-p1 -4.24074857 -1.491816683
#> TcYC6_0131900-RA-p1 -2.70178198 -0.743566839
#> TcYC6_0139040-RA-p1  9.83127459 -1.331652618
#> TcYC6_0130600-RA-p1 -2.86060511 -0.986892382
#> TcYC6_0136280-RA-p1 -3.01192850 -0.648472548
#> TcYC6_0136290-RA-p1 -0.29257355  6.982722605
#> TcYC6_0128520-RA-p1 -2.62652996 -0.354849679
#> TcYC6_0137910-RA-p1  7.11600858  0.220717825
#> TcYC6_0128530-RA-p1 -2.58205447 -0.994364448
#> TcYC6_0128540-RA-p1 -2.73186785 -0.594771355
#> TcYC6_0128550-RA-p1 -3.13219207 -0.866203254
#> TcYC6_0128560-RA-p1 -3.26675747 -1.063681507
#> TcYC6_0137920-RA-p1  9.23043274 -1.391784649
#> TcYC6_0135750-RA-p1 -0.29432530  6.986552884
#> TcYC6_0135760-RA-p1 -0.29432530  6.986552884
#> TcYC6_0135770-RA-p1 -0.30328126  6.988266447
#> TcYC6_0135780-RA-p1 -0.30328126  6.988266447
#> TcYC6_0135790-RA-p1 -0.29683556  6.948548427
#> TcYC6_0135800-RA-p1 -0.30112420  6.994196668
#> TcYC6_0135810-RA-p1 -0.29625052  6.984004160
#> TcYC6_0135820-RA-p1 -0.27723753  6.919307911
#> TcYC6_0135830-RA-p1 -0.29343197  6.954898341
#> TcYC6_0135840-RA-p1 -0.20556078  6.841972912
#> TcYC6_0128580-RA-p1 -3.96849138 -1.565676987
#> TcYC6_0128590-RA-p1 -2.82260097 -0.639715244
#> TcYC6_0128600-RA-p1 -2.57799014 -1.002367888
#> TcYC6_0128630-RA-p1 -2.21237198 -0.796247425
#> TcYC6_0135860-RA-p1 -0.29156693  6.939279742
#> TcYC6_0128650-RA-p1 -4.21594742 -1.536655526
#> TcYC6_0128660-RA-p1 -4.21909288 -1.445088508
#> TcYC6_0128670-RA-p1  0.09300741  5.877591087
#> TcYC6_0128680-RA-p1 -4.24219078 -1.422594421
#> TcYC6_0137990-RA-p1  9.71590122 -1.321368113
#> TcYC6_0035360-RA-p1  2.98265888  0.600515276
#> TcYC6_0128690-RA-p1 -3.23821514 -1.091901986
#> TcYC6_0138000-RA-p1  8.21501574 -0.606648195
#> TcYC6_0128900-RA-p1 -4.20748490 -1.410136969
#> TcYC6_0128910-RA-p1 -2.96174161 -1.058149031
#> TcYC6_0135890-RA-p1 -1.42973079 -0.170848487
#> TcYC6_0010720-RA-p1  2.98604046  0.633227587
#> TcYC6_0138010-RA-p1  9.72155023 -1.519394027
#> TcYC6_0002290-RA-p1  3.95759955  0.560377372
#> TcYC6_0129080-RA-p1 -2.87781305 -0.994135839
#> TcYC6_0129090-RA-p1 -2.85899987 -0.988609877
#> TcYC6_0129100-RA-p1 -2.85620535 -1.024154727
#> TcYC6_0129110-RA-p1 -2.83904121 -0.550227390
#> TcYC6_0129120-RA-p1 -2.83904121 -0.550227390
#> TcYC6_0129130-RA-p1 -2.81645933 -0.588662459
#> TcYC6_0129140-RA-p1 -2.81571648 -0.571238317
#> TcYC6_0129150-RA-p1 -2.82487089 -0.564620773
#> TcYC6_0129160-RA-p1 -2.85849782 -0.538038329
#> TcYC6_0129170-RA-p1 -2.86273188 -0.536128255
#> TcYC6_0135920-RA-p1 -2.67069970 -0.504064676
#> TcYC6_0138070-RA-p1  9.78017558 -1.370028212
#> TcYC6_0138080-RA-p1  9.72347196 -1.279335717
#> TcYC6_0138090-RA-p1  9.72347196 -1.279335717
#> TcYC6_0138100-RA-p1  9.72751995 -1.314359038
#> TcYC6_0138110-RA-p1  8.92954304 -0.486273133
#> TcYC6_0129180-RA-p1 -3.25744169 -1.116728296
#> TcYC6_0004460-RA-p1  3.95482308  0.547594828
#> TcYC6_0129190-RA-p1 -2.73028827 -0.592863631
#> TcYC6_0129200-RA-p1 -2.82644561 -0.909673526
#> TcYC6_0129210-RA-p1 -4.01862038 -1.344346969
#> TcYC6_0129220-RA-p1 -4.23828783 -1.442581501
#> TcYC6_0001540-RA-p1  3.88716785  0.566038023
#> TcYC6_0129310-RA-p1 -4.25783355 -1.461772196
#> TcYC6_0129330-RA-p1 -3.25330133 -1.018549476
#> TcYC6_0138220-RA-p1  9.24312916 -1.465614478
#> TcYC6_0129530-RA-p1 -3.32870520 -1.003878419
#> TcYC6_0136010-RA-p1 -0.30112420  6.994196668
#> TcYC6_0136020-RA-p1 -0.28929650  6.971923205
#> TcYC6_0129630-RA-p1 -2.88612544 -0.922575519
#> TcYC6_0129640-RA-p1 -2.76530756 -0.568052961
#> TcYC6_0129650-RA-p1 -4.26788609 -1.417831922
#> TcYC6_0129660-RA-p1  7.21197424  0.037156896
#> TcYC6_0129670-RA-p1  2.66397098 -0.794827905
#> TcYC6_0129680-RA-p1 -4.16389435 -1.549090421
#> TcYC6_0138240-RA-p1  9.42420713 -1.323972921
#> TcYC6_0129690-RA-p1 -2.78480680 -0.867378395
#> TcYC6_0129700-RA-p1 -3.13090903 -0.687016320
#> TcYC6_0138250-RA-p1 10.04238605 -1.318468462
#> TcYC6_0138260-RA-p1  8.18107210 -1.284314695
#> TcYC6_0129840-RA-p1 -3.75411041 -1.413440219
#> TcYC6_0129850-RA-p1 -4.14675175 -1.677595475
#> TcYC6_0129860-RA-p1 -4.15180897 -1.695834057
#> TcYC6_0129870-RA-p1 -4.08723325 -1.605151898
#> TcYC6_0129880-RA-p1 -4.05784451 -1.604143594
#> TcYC6_0138310-RA-p1  9.99456678 -1.509217407
#> TcYC6_0138320-RA-p1 10.28353970 -1.614471191
#> TcYC6_0136090-RA-p1 -0.30112420  6.994196668
#> TcYC6_0130120-RA-p1 -4.20323699 -1.408080399
#> TcYC6_0130130-RA-p1  6.99899151  0.128004418
#> TcYC6_0138450-RA-p1 10.31478230 -1.505017761
#> TcYC6_0130140-RA-p1 -2.96471670 -1.006340307
#> TcYC6_0130150-RA-p1 -2.96471670 -1.006340307
#> TcYC6_0136160-RA-p1 -0.30328126  6.988266447
#> TcYC6_0136170-RA-p1 -0.27246894  6.825140842
#> TcYC6_0136180-RA-p1 -0.29271725  6.987079041
#> TcYC6_0130200-RA-p1 -3.12738557 -0.882554778
#> TcYC6_0130220-RA-p1 -2.75087855 -0.583059837
#> TcYC6_0136200-RA-p1 -2.85563735 -0.920212631
#> TcYC6_0130230-RA-p1 -3.25610955 -1.022203693
#> TcYC6_0136210-RA-p1 -0.28327835  6.862115888
#> TcYC6_0003010-RA-p1  3.94379487  0.623296975
#> TcYC6_0130300-RA-p1 -3.07810915 -0.604101359
#> TcYC6_0130310-RA-p1 -3.09120553 -0.586575018
#> TcYC6_0130320-RA-p1 -3.03968781 -0.543436038
#> TcYC6_0136220-RA-p1 -1.55651146 -0.100765913
#> TcYC6_0130330-RA-p1  0.44714520  4.774290290
#> TcYC6_0130400-RA-p1 -2.87773456 -0.965846414
#> TcYC6_0130410-RA-p1 -2.87556689 -0.964307843
#> TcYC6_0136240-RA-p1 -2.62533763 -0.871911630
#> TcYC6_0138490-RA-p1  6.25390836 -1.118796841
#> TcYC6_0130420-RA-p1 -4.18844878 -1.509706363
#> TcYC6_0130430-RA-p1 -4.09505379 -1.379152849
#> TcYC6_0130440-RA-p1 -3.95223217 -1.643586285
#> TcYC6_0130450-RA-p1 -4.25344825 -1.460592643
#> TcYC6_0130550-RA-p1 -2.63596982 -0.951100037
#> TcYC6_0130560-RA-p1 -2.79666339 -0.583681031
#> TcYC6_0138560-RA-p1  9.63023321 -1.456061153
#> TcYC6_0130570-RA-p1 -3.13988836 -0.902425792
#> TcYC6_0130580-RA-p1 -2.75648161 -0.587088635
#> TcYC6_0130590-RA-p1 -2.74365472 -0.588115871
#> TcYC6_0130610-RA-p1 -0.25947649  6.702721440
#> TcYC6_0130620-RA-p1 -2.91766464 -1.045126754
#> TcYC6_0130630-RA-p1 -3.24135296 -0.915627768
#> TcYC6_0138570-RA-p1  9.57347090 -1.338774777
#> TcYC6_0130700-RA-p1 -4.03537316 -1.303150420
#> TcYC6_0130710-RA-p1 -4.22286193 -1.440931553
#> TcYC6_0130720-RA-p1 -4.17157801 -1.348738859
#> TcYC6_0130740-RA-p1 -4.15183743 -1.575975602
#> TcYC6_0130750-RA-p1 -4.21812717 -1.631666071
#> TcYC6_0130760-RA-p1 -4.14391027 -1.520730128
#> TcYC6_0029600-RA-p1  2.98817428  0.621600089
#> TcYC6_0138770-RA-p1  2.62331952 -0.568438354
#> TcYC6_0138780-RA-p1  2.61405030 -0.545792670
#> TcYC6_0131270-RA-p1 -3.01419180 -0.900665466
#> TcYC6_0136520-RA-p1 -3.99253782 -1.351101323
#> TcYC6_0136560-RA-p1 -0.30337254  6.977472811
#> TcYC6_0136570-RA-p1 -0.29343197  6.954898341
#> TcYC6_0131470-RA-p1 -0.25187100  6.662473793
#> TcYC6_0136580-RA-p1 -0.29774926  6.997399289
#> TcYC6_0136590-RA-p1 -0.29393912  6.978167123
#> TcYC6_0136600-RA-p1 -0.29774926  6.997399289
#> TcYC6_0136610-RA-p1 -0.30328126  6.988266447
#> TcYC6_0136620-RA-p1 -0.29755832  6.968994671
#> TcYC6_0131550-RA-p1 -4.34130411 -1.648127073
#> TcYC6_0131820-RA-p1 -4.09736530 -1.383593867
#> TcYC6_0021220-RA-p1  3.95759955  0.560377372
#> TcYC6_0139010-RA-p1  9.90272009 -1.408907993
#> TcYC6_0131860-RA-p1 -4.20578861 -1.415046118
#> TcYC6_0131870-RA-p1 -4.03533629 -1.583509351
#> TcYC6_0139050-RA-p1  9.72347196 -1.279335717
#> TcYC6_0136710-RA-p1 -0.29774926  6.997399289
#> TcYC6_0136720-RA-p1 -0.30181680  6.948855829
#> TcYC6_0136730-RA-p1 -0.29271459  6.947222593